+++ apps/backend/apps/admin/src/problem/services/problem.service.spec.ts
@@ -4,23 +4,11 @@ import { Test, type TestingModule } from '@nestjs/testing'
 import { Level } from '@generated'
 import { expect } from 'chai'
 import { spy, stub } from 'sinon'
-import {
-  EntityNotExistException,
-  UnprocessableDataException
-} from '@libs/exception'
+import { UnprocessableDataException } from '@libs/exception'
 import { PrismaService } from '@libs/prisma'
 import { S3MediaProvider, S3Provider } from '@admin/storage/s3.provider'
 import { StorageService } from '@admin/storage/storage.service'
 import {
-  exampleAssignment,
-  exampleAssignmentProblems,
-  exampleContest,
-  exampleContestProblems,
-  exampleOrderUpdatedAssignmentProblems,
-  exampleOrderUpdatedContestProblems,
-  exampleOrderUpdatedWorkbookProblems,
-  exampleWorkbook,
-  exampleWorkbookProblems,
   fileUploadInput,
   user,
   importedProblems,
@@ -32,7 +20,7 @@ import {
   testcaseInput,
   updateHistories
 } from '../mock/mock'
-import { FileService, ProblemService, TestcaseService } from './'
+import { FileService, ProblemService, TagService, TestcaseService } from './'
 
 /**
  * TODO: s3 관련 코드 재작성(수정) 필요
@@ -65,30 +53,6 @@ const db = {
     update: stub(),
     delete: stub()
   },
-  workbook: {
-    findFirstOrThrow: stub()
-  },
-  workbookProblem: {
-    findFirstOrThrow: stub(),
-    findMany: stub(),
-    update: stub()
-  },
-  contest: {
-    findFirstOrThrow: stub()
-  },
-  contestProblem: {
-    findFirstOrThrow: stub(),
-    findMany: stub(),
-    update: stub()
-  },
-  assignment: {
-    findFirstOrThrow: stub()
-  },
-  assignmentProblem: {
-    findFirstOrThrow: stub(),
-    findMany: stub(),
-    update: stub()
-  },
   $transaction: stub(),
   file: {
     deleteMany: stub()
@@ -115,6 +79,7 @@ describe('ProblemService', () => {
         { provide: PrismaService, useValue: db },
         StorageService,
         TestcaseService,
+        TagService,
         FileService,
         ConfigService,
         S3Provider,
@@ -343,379 +308,6 @@ describe('ProblemService', () => {
         user[0].id!
       )
       expect(result).to.deep.equal(problems[0])
-    })
-  })
-
-  describe('getWorkbookProblems', () => {
-    it('should return workbookProblems', async () => {
-      //given
-      db.workbook.findFirstOrThrow.resolves(exampleWorkbookProblems[0])
-      db.workbookProblem.findMany.resolves(exampleWorkbookProblems)
-      //when
-      const result = await service.getWorkbookProblems(1, 1)
-      //then
-      expect(result).to.deep.equals(exampleWorkbookProblems)
-    })
-
-    it('should handle NotExistError', async () => {
-      //given
-      db.workbook.findFirstOrThrow.rejects(
-        new EntityNotExistException('record not found')
-      )
-      // when & then
-      await expect(service.getWorkbookProblems(1, -1)).to.be.rejectedWith(
-        EntityNotExistException
-      )
-    })
-  })
-
-  describe('updateWorkbookProblemsOrder', () => {
-    beforeEach(() => {
-      // stub의 동작 초기화
-      db.workbookProblem.update.resetBehavior()
-      db.workbookProblem.findFirstOrThrow.resetBehavior()
-      db.workbookProblem.findMany.resetBehavior()
-    })
-    it('should return order-updated workbookProblems', async () => {
-      //given
-      const groupId = 1
-      const workbookId = 1
-      const orders = [2, 3, 4, 5, 6, 7, 8, 9, 10, 1]
-      const exampleWorkbookProblemsToBeUpdated =
-        exampleWorkbookProblems.toSorted((a, b) => a.problemId - b.problemId)
-      db.workbook.findFirstOrThrow.resolves(exampleWorkbook)
-      db.workbookProblem.findMany.resolves(exampleWorkbookProblems)
-
-      // update가 Promise.all로 실행되기 때문에 각 쿼리에 대한 모의 응답을 반환하도록 설정
-      for (let i = 0; i < 10; i++) {
-        const record = exampleWorkbookProblemsToBeUpdated[i]
-        const newOrder = orders.indexOf(record.problemId) + 1
-        db.workbookProblem.update
-          .withArgs({
-            where: {
-              // eslint-disable-next-line @typescript-eslint/naming-convention
-              workbookId_problemId: {
-                workbookId,
-                problemId: record.problemId
-              }
-            },
-            data: { order: newOrder }
-          })
-          .resolves(exampleOrderUpdatedWorkbookProblems[i])
-      }
-      db.$transaction.resolves(exampleOrderUpdatedWorkbookProblems)
-      //when
-      const result = await service.updateWorkbookProblemsOrder(
-        groupId,
-        workbookId,
-        [2, 3, 4, 5, 6, 7, 8, 9, 10, 1]
-      )
-      //then
-      expect(result).to.deep.equals(exampleOrderUpdatedWorkbookProblems)
-    })
-
-    it('should handle NotFound error', async () => {
-      //given
-      db.workbook.findFirstOrThrow.rejects(
-        new EntityNotExistException('record not found')
-      )
-      //when & then
-      await expect(
-        service.updateWorkbookProblemsOrder(
-          -1,
-          1,
-          [2, 3, 4, 5, 6, 7, 8, 9, 10, 1]
-        )
-      ).to.be.rejectedWith(EntityNotExistException)
-    })
-
-    it('should handle orders array length exception', async () => {
-      //given
-      db.workbook.findFirstOrThrow.resolves(exampleWorkbook)
-      db.workbookProblem.findMany.resolves(exampleWorkbookProblems)
-      //when & then
-      await expect(
-        service.updateWorkbookProblemsOrder(1, 1, [2, 3, 4, 5, 6, 7, 8, 9, 10])
-      ).to.be.rejectedWith(UnprocessableDataException)
-    })
-
-    it('should handle RecordNotFound error', async () => {
-      beforeEach(() => {
-        // stub의 동작 초기화
-        db.workbookProblem.update.resetBehavior()
-        db.workbookProblem.findFirstOrThrow.resetBehavior()
-        db.workbookProblem.findMany.resetBehavior()
-        db.$transaction.resetBehavior()
-      })
-      //given
-      db.workbook.findFirstOrThrow.resolves(exampleWorkbook)
-      db.workbookProblem.findMany.resolves(exampleWorkbookProblems)
-      db.workbookProblem.update.rejects(
-        new EntityNotExistException('record not found')
-      )
-      db.$transaction.rejects(new EntityNotExistException('record not found'))
-      //when & then
-      await expect(
-        service.updateWorkbookProblemsOrder(
-          1,
-          1,
-          [2, 3, 4, 5, 6, 7, 8, 9, 10, 1]
-        )
-      ).to.be.rejectedWith(EntityNotExistException)
-    })
-  })
-
-  describe('getContestProblems', () => {
-    it('should return ContestProblems', async () => {
-      //given
-      db.contest.findFirstOrThrow.resolves(exampleContest)
-      db.contestProblem.findMany.resolves(exampleContestProblems)
-      //when
-      const result = await service.getContestProblems(1)
-      //then
-      expect(result).to.deep.equals(exampleContestProblems)
-    })
-
-    it('should handle NotFoundError', async () => {
-      //given
-
-      db.contest.findFirstOrThrow.rejects(
-        new EntityNotExistException('record not found')
-      )
-      // when & then
-      await expect(service.getContestProblems(-1)).to.be.rejectedWith(
-        EntityNotExistException
-      )
-    })
-  })
-
-  describe('getAssignmentProblems', () => {
-    it('should return AssignmentProblems', async () => {
-      //given
-      db.assignment.findFirstOrThrow.resolves(exampleAssignment)
-      db.assignmentProblem.findMany.resolves(exampleAssignmentProblems)
-      //when
-      const result = await service.getAssignmentProblems(1, 1)
-      //then
-      expect(result).to.deep.equals(exampleAssignmentProblems)
-    })
-
-    it('should handle NotFoundError', async () => {
-      //given
-
-      db.assignment.findFirstOrThrow.rejects(
-        new EntityNotExistException('record not found')
-      )
-      // when & then
-      await expect(service.getAssignmentProblems(-1, 1)).to.be.rejectedWith(
-        EntityNotExistException
-      )
-    })
-  })
-
-  describe('updateAssignmentProblemsOrder', () => {
-    beforeEach(() => {
-      // 각 스텁의 동작 초기화
-      db.assignmentProblem.update.resetBehavior()
-      db.assignmentProblem.findFirstOrThrow.resetBehavior()
-      db.assignmentProblem.findMany.resetBehavior()
-      db.assignment.findFirstOrThrow.resetBehavior()
-    })
-    it('should return order-updated AssignmentProblems', async () => {
-      //given
-      const groupId = 1
-      const assignmentId = 1
-      const orders = [2, 3, 4, 5, 6, 7, 8, 9, 10, 1]
-      const exampleAssignmentProblemsToBeUpdated =
-        exampleAssignmentProblems.toSorted((a, b) => a.problemId - b.problemId)
-      db.assignment.findFirstOrThrow.resolves(exampleAssignment)
-      db.assignmentProblem.findMany.resolves(exampleAssignmentProblems)
-
-      // update가 Promise.all로 실행되기 때문에 각 쿼리에 대한 모의 응답을 반환하도록 설정
-      for (let i = 0; i < 10; i++) {
-        const record = exampleAssignmentProblemsToBeUpdated[i]
-        const newOrder = orders.indexOf(record.problemId) + 1
-        db.assignmentProblem.update
-          .withArgs({
-            where: {
-              // eslint-disable-next-line @typescript-eslint/naming-convention
-              assignmentId_problemId: {
-                assignmentId,
-                problemId: record.problemId
-              }
-            },
-            data: { order: newOrder }
-          })
-          .resolves(exampleOrderUpdatedAssignmentProblems[i])
-      }
-      db.$transaction.resolves(exampleOrderUpdatedAssignmentProblems)
-      //when
-      const result = await service.updateAssignmentProblemsOrder(
-        groupId,
-        assignmentId,
-        orders
-      )
-      //then
-      expect(result).to.deep.equals(exampleOrderUpdatedAssignmentProblems)
-    })
-
-    it('should handle NotFound error', async () => {
-      //given
-      db.assignment.findFirstOrThrow.rejects(
-        new EntityNotExistException('record not found')
-      )
-      //when & then
-      await expect(
-        service.updateAssignmentProblemsOrder(
-          -1,
-          1,
-          [2, 3, 4, 5, 6, 7, 8, 9, 10, 1]
-        )
-      ).to.be.rejectedWith(EntityNotExistException)
-    })
-
-    it('should handle orders array length exception', async () => {
-      //given
-      db.assignment.findFirstOrThrow.resolves(exampleAssignment)
-      db.assignmentProblem.findMany.resolves(exampleAssignmentProblems)
-      //when & then
-      await expect(
-        service.updateAssignmentProblemsOrder(
-          1,
-          1,
-          [2, 3, 4, 5, 6, 7, 8, 9, 10]
-        )
-      ).to.be.rejectedWith(UnprocessableDataException)
-    })
-
-    it('should handle RecordNotFound error', async () => {
-      beforeEach(() => {
-        // stub의 동작 초기화
-        db.workbookProblem.update.resetBehavior()
-        db.workbookProblem.findFirstOrThrow.resetBehavior()
-        db.workbookProblem.findMany.resetBehavior()
-        db.$transaction.resetBehavior()
-      })
-      //given
-      db.assignmentProblem.findMany.resolves(exampleAssignmentProblems)
-
-      db.assignmentProblem.update.rejects(
-        new EntityNotExistException('record not found')
-      )
-      db.$transaction.rejects(new EntityNotExistException('record not found'))
-      //when & then
-      await expect(
-        service.updateAssignmentProblemsOrder(
-          1,
-          1,
-          [2, 3, 4, 5, 6, 7, 8, 9, 10, 1]
-        )
-      ).to.be.rejectedWith(EntityNotExistException)
-    })
-  })
-
-  describe('updateContestProblemsOrder', () => {
-    beforeEach(() => {
-      // 각 스텁의 동작 초기화
-      db.contestProblem.update.resetBehavior()
-      db.contestProblem.findFirstOrThrow.resetBehavior()
-      db.contestProblem.findMany.resetBehavior()
-      db.contest.findFirstOrThrow.resetBehavior()
-    })
-    it('should return order-updated ContestProblems', async () => {
-      //given
-      const contestId = 1
-      const orders = [2, 3, 4, 5, 6, 7, 8, 9, 10, 1]
-      const exampleContestProblemsToBeUpdated = exampleContestProblems.toSorted(
-        (a, b) => a.problemId - b.problemId
-      )
-      db.contest.findFirstOrThrow.resolves(exampleContest)
-      db.contestProblem.findMany.resolves(exampleContestProblems)
-
-      // update가 Promise.all로 실행되기 때문에 각 쿼리에 대한 모의 응답을 반환하도록 설정
-      for (let i = 0; i < 10; i++) {
-        const record = exampleContestProblemsToBeUpdated[i]
-        const newOrder = orders.indexOf(record.problemId) + 1
-        db.contestProblem.update
-          .withArgs({
-            where: {
-              // eslint-disable-next-line @typescript-eslint/naming-convention
-              contestId_problemId: {
-                contestId,
-                problemId: record.problemId
-              }
-            },
-            data: { order: newOrder }
-          })
-          .resolves(exampleOrderUpdatedContestProblems[i])
-      }
-      db.$transaction.resolves(exampleOrderUpdatedContestProblems)
-      //when
-      const result = await service.updateContestProblemsOrder(contestId, orders)
-      //then
-      expect(result).to.deep.equals(exampleOrderUpdatedContestProblems)
-    })
-
-    it('should Error when orders length is not same with problems length', async () => {
-      //given
-      db.contest.findFirstOrThrow.resolves(exampleContest)
-      db.contestProblem.findMany.resolves(exampleContestProblems)
-      //when & then
-      await expect(
-        service.updateContestProblemsOrder(1, [2, 3, 4, 5, 6, 7, 8, 9, 10])
-      ).to.be.rejectedWith(UnprocessableDataException)
-    })
-
-    it('should Error when orders dont have full imported problems', async () => {
-      //given
-      db.contest.findFirstOrThrow.resolves(exampleContest)
-      db.contestProblem.findMany.resolves(exampleContestProblems)
-      //when & then
-      await expect(
-        service.updateContestProblemsOrder(1, [2, 2, 4, 5, 6, 7, 8, 9, 10, 1])
-      ).to.be.rejectedWith(UnprocessableDataException)
-    })
-
-    it('should handle NotFound error', async () => {
-      //given
-      db.contest.findFirstOrThrow.rejects(
-        new EntityNotExistException('record not found')
-      )
-      //when & then
-      await expect(
-        service.updateContestProblemsOrder(-1, [2, 3, 4, 5, 6, 7, 8, 9, 10, 1])
-      ).to.be.rejectedWith(EntityNotExistException)
-    })
-
-    it('should handle orders array length exception', async () => {
-      //given
-      db.contest.findFirstOrThrow.resolves(exampleContest)
-      db.contestProblem.findMany.resolves(exampleContestProblems)
-      //when & then
-      await expect(
-        service.updateContestProblemsOrder(1, [2, 3, 4, 5, 6, 7, 8, 9, 10])
-      ).to.be.rejectedWith(UnprocessableDataException)
-    })
-
-    it('should handle RecordNotFound error', async () => {
-      beforeEach(() => {
-        // stub의 동작 초기화
-        db.workbookProblem.update.resetBehavior()
-        db.workbookProblem.findFirstOrThrow.resetBehavior()
-        db.workbookProblem.findMany.resetBehavior()
-        db.$transaction.resetBehavior()
-      })
-      //given
-      db.contestProblem.findMany.resolves(exampleContestProblems)
-
-      db.contestProblem.update.rejects(
-        new EntityNotExistException('record not found')
-      )
-      db.$transaction.rejects(new EntityNotExistException('record not found'))
-      //when & then
-      await expect(
-        service.updateContestProblemsOrder(1, [2, 3, 4, 5, 6, 7, 8, 9, 10, 1])
-      ).to.be.rejectedWith(EntityNotExistException)
     })
   })
 })
